+++ core/source.ts
@@ -670,7 +670,7 @@ export function getUrl(
   const { data } = page;
   let url = data.url as
     | string
-    | ((page: Page) => string | false)
+    | ((page: Page, defaultUrl: string) => string | false)
     | false
     | undefined;
 
@@ -679,7 +679,7 @@ export function getUrl(
   }
 
   if (typeof url === "function") {
-    url = url(page);
+    url = url(page, getDefaultUrl(page, parentPath, prettyUrls));
   }
 
   if (typeof url === "string") {
@@ -704,8 +704,16 @@ export function getUrl(
     );
   }
 
+  return getDefaultUrl(page, parentPath, prettyUrls);
+}
+
+function getDefaultUrl(
+  page: Page,
+  parentPath: string,
+  prettyUrls: boolean,
+): string {
   // Calculate the URL from the path
-  url = posix.join(parentPath, page.src.slug);
+  const url = posix.join(parentPath, page.src.slug);
 
   if (page.src.asset) {
     return url + page.src.ext;